\name{BayesQArt-package}
\alias{BayesQArt-package}
\alias{BayesQArt}
\docType{package}
\title{
\packageTitle{BayesQArt}
}
\description{
\packageDescription{BayesQArt}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BayesQArt}
\packageIndices{BayesQArt}
}
\author{
\packageAuthor{BayesQArt}

Maintainer: \packageMaintainer{BayesQArt}
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
}
