% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmad.R
\name{wmad}
\alias{wmad}
\title{A function to calculate weigthed mean absolute deviation.}
\usage{
wmad(y_true, y_hat, quant)
}
\arguments{
\item{y_true}{actual response,}

\item{y_hat}{conditional quantile prediction,}

\item{quant}{quantile value.}
}
\value{
weighted mean absoulute deviation, \code{wmad}.
}
\description{
A function to calculate weigthed mean absolute deviation.
}

