\name{BayesQArt-package}
\alias{BayesQArt-package}
\alias{BayesQArt}
\docType{package}
\title{
\packageTitle{BayesQArt}
}
\description{
\packageDescription{BayesQArt}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BayesQArt}
\packageIndices{BayesQArt}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{BayesQArt}

Maintainer: \packageMaintainer{BayesQArt}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
